// application/typescript/library/amplify/backend/function/eventLogStreamProcessor/lib/library/codegen/API.ts
export type PoolPlayer = {
  __typename: "PoolPlayer",
  id: string,
  userWallet: UserWallet,
  status: PlayerStatus,
  pool: Pool,
  createdAt: string,
  updatedAt: string,
  poolPlayersId?: string | null,
  poolPlayerUserWalletId: string,
};

export type UserWallet = {
  __typename: "UserWallet",
  id: string,
  wallet: string,
  nickname?: string | null,
  chatlogo?: string | null,
  brands?: ModelBrandConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelBrandConnection = {
  __typename: "ModelBrandConnection",
  items:  Array<Brand | null >,
  nextToken?: string | null,
};

export type Brand = {
  __typename: "Brand",
  id: string,
  wallet: string,
  name: string,
  rating: number,
  logo?: string | null,
  completedPools: number,
  failedPools: number,
  createdAt: string,
  updatedAt: string,
  userWalletBrandsId?: string | null,
};

export type Pool = {
  __typename: "Pool",
  id: string,
  poolId: number,
  poolTitle: string,
  poolCategory: PoolCategory,
  poolCreator: UserWallet,
  poolType: poolType,
  poolStatus: string,
  poolEntryFee: string,
  poolTotal: string,
  poolWinningPayout: string,
  allowPlayerLeave: boolean,
  apiPoolAttributes?: ApiPoolAttributes | null,
  lotteryPoolAttributes?: LotteryPoolAttributes | null,
  players?: ModelPoolPlayerConnection | null,
  createdAt: string,
  updatedAt: string,
  poolPoolCreatorId: string,
  poolApiPoolAttributesId?: string | null,
  poolLotteryPoolAttributesId?: string | null,
};

export enum PoolCategory {
  other = "other",
  sports = "sports",
  esports = "esports",
  gaming = "gaming",
  politics = "politics",
  financial = "financial",
  random_winners = "random_winners",
}


export enum poolType {
  lottery = "lottery",
  manual = "manual",
  api = "api",
  custom_builder = "custom_builder",
}


export type ApiPoolAttributes = {
  __typename: "ApiPoolAttributes",
  id: string,
  apiKey: string,
  lockFundsDatetime: string,
  executeWinnerDatetime: string,
  apiUrlForResults: string,
  verifiedUrlSchema: boolean,
  poolCreatorPercentFeeToWithold: number,
  apiWinnerOptionEntries: string,
  apiwinnerResult: string,
  poolApiDefaultSchema: string,
  poolApiCustomSchema?: string | null,
  pool?: Pool | null,
  createdAt: string,
  updatedAt: string,
  apiPoolAttributesPoolId?: string | null,
};

export type LotteryPoolAttributes = {
  __typename: "LotteryPoolAttributes",
  id: string,
  auditRecordDrawId?: string | null,
  isAuditEnabled: boolean,
  randomOrgUrlForResults?: string | null,
  pool: Pool,
  createdAt: string,
  updatedAt: string,
  lotteryPoolAttributesPoolId: string,
};

export type ModelPoolPlayerConnection = {
  __typename: "ModelPoolPlayerConnection",
  items:  Array<PoolPlayer | null >,
  nextToken?: string | null,
};
// ...
export type CreatePoolInput = {
  id?: string | null,
  poolId: number,
  poolTitle: string,
  poolCategory: PoolCategory,
  poolType: poolType,
  poolStatus: string,
  poolEntryFee: string,
  poolTotal: string,
  poolWinningPayout: string,
  allowPlayerLeave: boolean,
  poolPoolCreatorId: string,
  poolApiPoolAttributesId?: string | null,
  poolLotteryPoolAttributesId?: string | null,
};
// ...
export type UpdatePoolInput = {
  id: string,
  poolId?: number | null,
  poolTitle?: string | null,
  poolCategory?: PoolCategory | null,
  poolType?: poolType | null,
  poolStatus?: string | null,
  poolEntryFee?: string | null,
  poolTotal?: string | null,
  poolWinningPayout?: string | null,
  allowPlayerLeave?: boolean | null,
  poolPoolCreatorId?: string | null,
  poolApiPoolAttributesId?: string | null,
  poolLotteryPoolAttributesId?: string | null,
};
// ...
export type SearchableIDFilterInput = {
  ne?: string | null,
  gt?: string | null,
  lt?: string | null,
  gte?: string | null,
  lte?: string | null,
  eq?: string | null,
  match?: string | null,
  matchPhrase?: string | null,
  matchPhrasePrefix?: string | null,
  multiMatch?: string | null,
  exists?: boolean | null,
  wildcard?: string | null,
  regexp?: string | null,
  range?: Array< string | null > | null,
};

export type SearchableStringFilterInput = {
  ne?: string | null,
  gt?: string | null,
  lt?: string | null,
  gte?: string | null,
  lte?: string | null,
  eq?: string | null,
  match?: string | null,
  matchPhrase?: string | null,
  matchPhrasePrefix?: string | null,
  multiMatch?: string | null,
  exists?: boolean | null,
  wildcard?: string | null,
  regexp?: string | null,
  range?: Array< string | null > | null,
};
// ...
export type SearchableAggregateResult = {
  __typename: "SearchableAggregateResult",
  name: string,
  result?: SearchableAggregateGenericResult | null,
};

export type SearchableAggregateGenericResult = SearchableAggregateScalarResult | SearchableAggregateBucketResult


export type SearchableAggregateScalarResult = {
  __typename: "SearchableAggregateScalarResult",
  value: number,
};

export type SearchableAggregateBucketResult = {
  __typename: "SearchableAggregateBucketResult",
  buckets?:  Array<SearchableAggregateBucketResultItem | null > | null,
};

export type SearchableAggregateBucketResultItem = {
  __typename: "SearchableAggregateBucketResultItem",
  key: string,
  doc_count: number,
};
// ...
export type SearchableBooleanFilterInput = {
  eq?: boolean | null,
  ne?: boolean | null,
};
// ...
export type SearchableIntFilterInput = {
  ne?: number | null,
  gt?: number | null,
  lt?: number | null,
  gte?: number | null,
  lte?: number | null,
  eq?: number | null,
  range?: Array< number | null > | null,
};
// ...
export type SearchablePoolFilterInput = {
  id?: SearchableIDFilterInput | null,
  poolId?: SearchableIntFilterInput | null,
  poolTitle?: SearchableStringFilterInput | null,
  poolStatus?: SearchableStringFilterInput | null,
  poolEntryFee?: SearchableStringFilterInput | null,
  poolTotal?: SearchableStringFilterInput | null,
  poolWinningPayout?: SearchableStringFilterInput | null,
  allowPlayerLeave?: SearchableBooleanFilterInput | null,
  createdAt?: SearchableStringFilterInput | null,
  updatedAt?: SearchableStringFilterInput | null,
  poolPoolCreatorId?: SearchableIDFilterInput | null,
  poolApiPoolAttributesId?: SearchableIDFilterInput | null,
  poolLotteryPoolAttributesId?: SearchableIDFilterInput | null,
  poolCategory?: SearchableStringFilterInput | null,
  poolType?: SearchableStringFilterInput | null,
  and?: Array< SearchablePoolFilterInput | null > | null,
  or?: Array< SearchablePoolFilterInput | null > | null,
  not?: SearchablePoolFilterInput | null,
};
// ...
export type SearchablePoolConnection = {
  __typename: "SearchablePoolConnection",
  items:  Array<Pool | null >,
  nextToken?: string | null,
  total?: number | null,
  aggregateItems:  Array<SearchableAggregateResult | null >,
};
// ...
export type CreatePoolMutation = {
  createPool?:  {
    __typename: "Pool",
    id: string,
    poolId: number,
    poolTitle: string,
    poolCategory: PoolCategory,
    poolCreator:  {
      __typename: "UserWallet",
      id: string,
      wallet: string,
      nickname?: string | null,
      chatlogo?: string | null,
      createdAt: string,
      updatedAt: string,
    },
    poolType: poolType,
    poolStatus: string,
    poolEntryFee: string,
    poolTotal: string,
    poolWinningPayout: string,
    allowPlayerLeave: boolean,
    apiPoolAttributes?:  {
      __typename: "ApiPoolAttributes",
      id: string,
      apiKey: string,
      lockFundsDatetime: string,
      executeWinnerDatetime: string,
      apiUrlForResults: string,
      verifiedUrlSchema: boolean,
      poolCreatorPercentFeeToWithold: number,
      apiWinnerOptionEntries: string,
      apiwinnerResult: string,
      poolApiDefaultSchema: string,
      poolApiCustomSchema?: string | null,
      createdAt: string,
      updatedAt: string,
      apiPoolAttributesPoolId?: string | null,
    } | null,
    lotteryPoolAttributes?:  {
      __typename: "LotteryPoolAttributes",
      id: string,
      auditRecordDrawId?: string | null,
      isAuditEnabled: boolean,
      randomOrgUrlForResults?: string | null,
      createdAt: string,
      updatedAt: string,
      lotteryPoolAttributesPoolId: string,
    } | null,
    players?:  {
      __typename: "ModelPoolPlayerConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    poolPoolCreatorId: string,
    poolApiPoolAttributesId?: string | null,
    poolLotteryPoolAttributesId?: string | null,
  } | null,
};

export type UpdatePoolMutationVariables = {
  input: UpdatePoolInput,
  condition?: ModelPoolConditionInput | null,
};

export type UpdatePoolMutation = {
  updatePool?:  {
    __typename: "Pool",
    id: string,
    poolId: number,
    poolTitle: string,
    poolCategory: PoolCategory,
    poolCreator:  {
      __typename: "UserWallet",
      id: string,
      wallet: string,
      nickname?: string | null,
      chatlogo?: string | null,
      createdAt: string,
      updatedAt: string,
    },
    poolType: poolType,
    poolStatus: string,
    poolEntryFee: string,
    poolTotal: string,
    poolWinningPayout: string,
    allowPlayerLeave: boolean,
    apiPoolAttributes?:  {
      __typename: "ApiPoolAttributes",
      id: string,
      apiKey: string,
      lockFundsDatetime: string,
      executeWinnerDatetime: string,
      apiUrlForResults: string,
      verifiedUrlSchema: boolean,
      poolCreatorPercentFeeToWithold: number,
      apiWinnerOptionEntries: string,
      apiwinnerResult: string,
      poolApiDefaultSchema: string,
      poolApiCustomSchema?: string | null,
      createdAt: string,
      updatedAt: string,
      apiPoolAttributesPoolId?: string | null,
    } | null,
    lotteryPoolAttributes?:  {
      __typename: "LotteryPoolAttributes",
      id: string,
      auditRecordDrawId?: string | null,
      isAuditEnabled: boolean,
      randomOrgUrlForResults?: string | null,
      createdAt: string,
      updatedAt: string,
      lotteryPoolAttributesPoolId: string,
    } | null,
    players?:  {
      __typename: "ModelPoolPlayerConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    poolPoolCreatorId: string,
    poolApiPoolAttributesId?: string | null,
    poolLotteryPoolAttributesId?: string | null,
  } | null,
};
// ...
export type SearchPoolsQuery = {
  searchPools?:  {
    __typename: "SearchablePoolConnection",
    items:  Array< {
      __typename: "Pool",
      id: string,
      poolId: number,
      poolTitle: string,
      poolCategory: PoolCategory,
      poolType: poolType,
      poolStatus: string,
      poolEntryFee: string,
      poolTotal: string,
      poolWinningPayout: string,
      allowPlayerLeave: boolean,
      createdAt: string,
      updatedAt: string,
      poolPoolCreatorId: string,
      poolApiPoolAttributesId?: string | null,
      poolLotteryPoolAttributesId?: string | null,
    } | null >,
    nextToken?: string | null,
    total?: number | null,
    aggregateItems:  Array< {
      __typename: "SearchableAggregateResult",
      name: string,
      result: ( {
          __typename: "SearchableAggregateScalarResult",
          value: number,
        } | {
          __typename: "SearchableAggregateBucketResult",
          buckets?:  Array< {
            __typename: string,
            key: string,
            doc_count: number,
          } | null > | null,
        }
      ) | null,
    } | null >,
  } | null,
};